refactored

+++ FilleUploadCore/UploadManagers/UploadServiceBase.cs
@@ -61,10 +61,10 @@ namespace FilleUploadCore.UploadManagers
 
         public long? UserId { get; set; }
 
-        public int? ApprovalConfigId { get; set; }
+        public long? ApprovalConfigId { get; set; }
 
         public string UserName { get; set; }
 
-        public int? BusinessId { get; set; }
+        public long? BusinessId { get; set; }
     }
 }
